mgurdy/requirements.txt
aniso8601==1.2.1
appdirs==1.4.3
cffi==1.10.0
click==6.7
decorator==4.0.11
flake8==3.3.0
Flask==0.12.2
Flask-RESTful==0.3.6
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
olefile==0.44
packaging==16.8
pbr==3.0.0
peewee==2.10.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
pyalsaaudio==0.8.4
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.0.7
pytest-timeout==1.2.0
python-dateutil==2.6.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
